fixed some libft stuff

+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: gjakobss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/15 11:05:01 by gjakobss          #+#    #+#             */
-/*   Updated: 2021/02/19 16:48:05 by gjakobss         ###   ########.fr       */
+/*   Updated: 2021/11/09 11:21:25 by malmeida         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 size_t	check_len(char const *s, size_t len)
 {
-	if (len > ft_strlen(s))
+	if ((int)len > ft_strlen(s))
 		return (ft_strlen(s));
 	return (len);
 }
@@ -28,7 +28,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	i = 0;
 	if (!s)
 		return (NULL);
-	if (start > ft_strlen(s))
+	if ((int)start > ft_strlen(s))
 	{
 		ptr = malloc(sizeof(char));
 		if (ptr == NULL)
@@ -39,7 +39,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	ptr = malloc((check_len(s, len) + 1) * sizeof(char));
 	if (ptr == NULL)
 		return (NULL);
-	while (i < len && start + i < ft_strlen(s))
+	while (i < len && (int)(start + i) < ft_strlen(s))
 	{
 		ptr[i] = s[start + i];
 		i++;
